review methods mailer
