+++ src/Templates/register/userExist.twig
@@ -3,8 +3,17 @@
 {% block title %} Utilisateur existant {% endblock %}
 
 {% block body %}
+
+{% if email is not null %}
 <h1>Utilisateur déjà enregistré</h1>
 <p>Nous avons déjà un utilisateur enregistré avec l'adresse mail : {{ email }}</p>
+{% endif %}
+
+{% if pseudonym is not null %}
+<h1>Utilisateur déjà enregistré</h1>
+<p>Ce pseudonym est déjà utilisé : {{ pseudonym }}</p>
+{% endif %}
+
 {% block content %}
 {# {% include "post/index.twig" %} #}
 {% endblock content %}
